Initial move over from core (#1)
There's a bunch of general math utilities in core that Mesh will depend on, core depends on. More, there's a bunch of stuff that depends on core in part just because it needs these generic math utilities. So this seemed like a good first thing to break out.

This PR is really all structural, there are no interesting changes in the actual implementations.